Fixes based on feedback.
