+++ CS483/CS483/Kartaclysm/StateMachine/Gameplay/StateMainMenu.h
@@ -13,7 +13,6 @@
 #include "ComponentPerspectiveCamera.h"
 
 #include "GameplayState.h"
-#include "GameplayStatesEnum.h"
 #include "EventManager.h"
 #include "KeyboardInputBuffer.h"
