+++ CS483/CS483/Kartaclysm/StateMachine/Gameplay/StatePlayerSelectionMenu.h
@@ -17,7 +17,6 @@
 #include "Component3DModel.h"
 
 #include "GameplayState.h"
-#include "GameplayStatesEnum.h"
 #include "EventManager.h"
 #include "KeyboardInputBuffer.h"
