+++ CS483/CS483/Kartaclysm/StateMachine/Gameplay/StateRaceCompleteMenu.h
@@ -14,7 +14,6 @@
 #include "ComponentTextbox.h"
 
 #include "GameplayState.h"
-#include "GameplayStatesEnum.h"
 #include "EventManager.h"
 #include "KeyboardInputBuffer.h"
