+++ CS483/CS483/Kartaclysm/StateMachine/Gameplay/StateRacing.h
@@ -48,7 +48,6 @@
 
 #include "Common.h"
 #include "GameplayState.h"
-#include "GameplayStatesEnum.h"
 #include "EventManager.h"
 
 #include "ComponentRacer.h"
