+++ CS483/CS483/Kartaclysm/StateMachine/Gameplay/StateTrackSelectionMenu.h
@@ -15,7 +15,6 @@
 #include "ComponentTextbox.h"
 
 #include "GameplayState.h"
-#include "GameplayStatesEnum.h"
 #include "EventManager.h"
 #include "KeyboardInputBuffer.h"
